modif formulaire création bien 18-06-22

+++ src/main/java/com/example/rdc_touristique/rest_api/PersonneController.java
@@ -87,7 +87,7 @@ public class PersonneController extends AbstratCrudController<PersonneSimpleDTO,
     }
 
     @GetMapping("/nbcompte")
-    public int nbCompte(){
+    public long nbCompte(){
         return ((PersonneService)service).nbCompte();
     }
